+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: famartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/23 16:15:53 by famartin          #+#    #+#             */
-/*   Updated: 2018/02/27 16:01:20 by famartin         ###   ########.fr       */
+/*   Updated: 2018/03/07 13:32:57 by famartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int				i;
-	unsigned char	*ptr1;
-	unsigned char	*ptr2;
+	size_t				i;
+	unsigned char		*ptr1;
+	unsigned char		*ptr2;
 
 	i = 0;
 	ptr1 = (unsigned char *)s1;
