+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: famartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/23 16:14:43 by famartin          #+#    #+#             */
-/*   Updated: 2018/02/26 15:40:28 by famartin         ###   ########.fr       */
+/*   Updated: 2018/03/07 13:31:44 by famartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
-	int		i;
-	char	*dst_ptr;
-	char	*src_ptr;
+	size_t		i;
+	char		*dst_ptr;
+	char		*src_ptr;
 
 	i = 0;
 	dst_ptr = dst;
