+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: famartin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/23 16:16:52 by famartin          #+#    #+#             */
-/*   Updated: 2018/03/01 18:36:37 by famartin         ###   ########.fr       */
+/*   Updated: 2018/03/07 13:30:21 by famartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *restrict s1, const char *restrict s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = (int)ft_strlen(s1);
